Create Godot project
I don’t know if Godot is the best engine to use for this game, but Godot
is the engine that I’m the most familiar with, so it’s the one that I’m
using.

Instead of using the default icon.svg file that comes with Godot
projects, I decided to include a 1 pixel image as the icon. Using the 1
pixel placeholder icon has multiple advantages:

1. It’s smaller.

2. It doesn’t require any attribution. The default Godot project icon
contains a Godot logo. The Godot logo is available under 🅭🅯4.0 [1]. It’s
easier to create a 1 pixel icon for my game than it is to comply with a
license.

[1]: <https://github.com/godotengine/godot/blob/4.2.2-stable/LOGO_LICENSE.txt>