Always use a 16:9 aspect ratio
The main motivation behind this change is to make it easier to create
future commits. Those future commits will add additional screens to the
game. If I know the aspect ratio of the game, then I will be able to
make those additional screens more easily. It’s easier to only worry
about a single aspect ratio. Making screens that work in almost any
aspect ratio is much harder.

I chose the resolution 1920x1080 in particular because it’s a common
resolution for video, and future commits will add videos that play while
you are playing the game.