Create legal notices screen
If you want to distribute a compiled Godot game, then you have to
include some Godot-related legal notices in your distribution [1].
Including a legal notices screen inside the game itself helps ensure
that everyone why receives a copy of the game also receives a copy of
the required legal notices.

I also decided to include the list of Godot project donors. I don’t
think that that list is legally required, but it was easy to include and
it make my user interface more similar to the Godot editor’s user
interface. Specifically, I modeled my legal notices screen after the
Godot editor’s Help → About Godot… menu.

This commit adds a pointless TabContainer named ProjectSelector. The
ProjectSelector TabContainer is pointless because it only has one tab.
The only reason that I added it in this commit is to make it easier to
create a future commit. That future commit is going to add a second tab
to that tab container.

[1]: <https://docs.godotengine.org/en/4.2/about/complying_with_licenses.html>