Make legal notice screen accessible
Before this change, the legal notices screen existed, but the only way
to access it was to run the game in the Godot editor and use the Run
Current Scene button. If you were playing the game normally, then there
was no way to view the legal notices screen. This commit fixes that
problem.