Add legal notice for Type That Tune’s source code
This change will make it easier for me to incorporate third-party assets
into my code. For example, in a future commit, I’m going to add a font
to the project. The font requires that you include a legal notice when
you distribute the font. The changes in this commit will make it easy to
add a legal notice for that font and for any other third-party assets
that I end up incorporating into my game.