Download media for first level
Now that the build tool downloads a video and converts it to a format
that can be used by Godot, I can start working on the first level of the
game.

The build tool caches any videos that have been downloaded via yt-dlp.
When debugging the build tool, it might make sense to run the build tool
over and over again. If the build tool tries to download a video over
and over again, then the build tool might get rate limited. Caching the
downloaded videos will hopefully prevent that from happening.