Create BaseLevel
The idea is that all levels will inherit from BaseLevel. Also, I decided
to put a limitation on the player. The player will not be allowed to
enter incorrect characters. The goal of this limitation is to make the
game easier. We’ll see if making the game easier makes it more fun for
new players. In the future, I want to add a difficult system. On harder
difficulties, players will be able to enter incorrect characters.