Allow text queue to loop
This will be useful for segments where the player has to type in the
same title or titles over and over again. I also added a mechanism to
disable queue looping. I want to be able to disable queue looping so
that I can have players type in the name of a song one time in order to
start a level.