Create first part of level 1
I tried to make the first part of this level slow on purpose in order to
ease players into the game. That’s why I chose Clown’s Tenth Anniversary
in particular for the first level.

I’m taking a break before finishing level 1 in order to fix a bug with
BaseLevel.