Prevent player from not being able to type
Before this commit, if a player had partially typed in some text, and
then the text queue was changed, then the player would be softlocked.
There would already be text in the UnserInput text box, but that text
would be incorrect because it was validated for the previous song title
(not the current one). Players wouldn’t be able to add additional
characters because the previous characters were already wrong. Even if a
player was able to enter additional characters, they wouldn’t be able to
submit their text because what they had typed in wouldn’t have matched
the correct answer.