Add pause menu
One of the reasons why I’m adding this now is to make the game easier to
debug. In a future commit, I’m going to add an option to the pause menu
that will allow you to skip ahead in the level. That option will make it
easier for me to fix a problem where a specific song title appears too
early.