Switch to _unhandled_key_input()
The only difference between _unhandled_input() and
_unhandled_key_input() is that _unhandled_key_input() only receives
InputEvents when a user presses keys. It won’t receive InputEvents when
a player uses a controller or a mouse.

There’s no reason for us to ever receive InputEvents from controllers or
mice. Type That Tune is a typing game, so we only care about InputEvents
from keyboards. I haven’t tested this, but I suspect that only
processing input from keyboards makes the code slightly faster. That’s
the main motivation behind this change.