Change input required to skip the tutorial
Before this change, if a player tried to pause the game during the
tutorial, then the game would skip the tutorial. This commit makes it so
that you have to press Ctrl+Space in order to skip the tutorial. I chose
Ctrl+Space because it’s easy to press and isn’t likely to be pressed by
accident.

It’s going to be important to be able to pause during the tutorial
because I’m going to add a part to the tutorial that teaches players how
to pause.