Create level for The Glorious Octagon of Destiny
I chose The Glorious Octagon of Destiny for a very specific reason.
Every segment in Type That Tune’s first level has a single item
text_queue. That was on purpose. I wanted to make the first level very
simple in order to ease people into the game. For the second level, I
wanted to choose a song that had a single segment that had multiple
items in its text_queue. The Glorious Octagon of Destiny has a single
segment with multiple items in its text_queue. Additionally, The
Glorious Octagon of Destiny is not too long. It’s only eight minutes
long which is shorter than many YTPMV medleys.

Also, I added some code to the build tool that limits the maximum width
of The Glorious Octagon of Destiny video to 1920. The version of the
video that’s available on YouTube has a maximum resolution of 8K. I
doubt that my machine or many other machines could handle 8K libtheora
video playback. Plus, I doubt that many people have 8K screens. My
screen is only 1920x1200.

Closes #13.