Fix potential crash when shortening the text_queue
Before this change, the game was likely to crash if a level switched
from a longer text_queue to a shorter text_queue. For example, I’m
working on a future commit that will add a level for The Glorious
Octagon of Destiny. The Glorious Octagon of Destiny has a “zoomzoom”
segment that contains 11 different songs. After that segment is the
credits, and the credits contain a single song [1]. When that future
commit switched from the text_queue for the zoomzoom segment to the
text_queue for the credits segment, the game would crash. This commit
prevents that problem from happening.

[1]: <https://youtu.be/5VWF-TaIheE>