Add BaseLevel.initial_text_queue property
Before this change, I had to fight with the Godot editor in order to get
it to not break The Glorious Octagon of Destiny’s level. Whenever I
opened The Glorious Octagon of Destiny’s level, it would set its
initial_text_queue property to []. It would do that because the
AnimationPlayer’s first key value for that property was []. Adding an
initial_text_queue property allows me to sidestep that problem. The
initial_text_queue property isn’t controlled by any AnimationPlayers, so
scrubbing AnimationPlayers in the editor shouldn’t change the
initial_text_queue property.