Make sure that videos fill the screen
Before this change, Ronald McDonald Insanity would be small and appear
in the top left corner of the screen. This commit fixes that problem.

Instead of having developers set the ratio property of the AspectRatio
container, I gave the BaseLevel Node a separate property called
aspect_ratio. Setting BaseLevel.aspect_ratio will eventually cause
BaseLevel.get_node("AspectRatioContainer").ratio to change. I did it
like this because AspectRatio.ratio is a float. When setting a float
property in the Godot editor, it only lets you enter a number with a
decimal point. I don’t want to enter a decimal number. I want to enter a
fraction like 16:9 or 4:3. Whenever I think about aspect ratios, I
always think of them as fractions. It’s going to be less error prone for
me to enter fractions because that’s what I’m used to. It’s also going
to be more precise. When you write 16:9 as a decimal number, you have to
round it because you can’t write an infinite number of digits. That’s
why I created the BaseLevel.aspect_ratio property. The
BaseLevel.aspect_ratio property allows me to enter an exact fraction
rather than an approximate decimal number.