Create level for the 1st Ran Ran Ru collaboration
I had to romanize the name of one of the songs (君にBUMP). I wasn’t sure
what words should be capitalized and what words shouldn’t be
capitalized. I found this Stack Exchange answer [1]. That Stack Exchange
answer mentioned that there was a guide for romanization that the
Library of Congress uses [2]. That guide says

> Capitalize all proper names following the rules below. Never
> capitalize particles.

It also says

> Capitalize the first word of names of works such as documents,
> publications, artworks, etc.

Based on that information, I believe that 君にBUMP romanizes to “Kimi ni
BUMP”. The word “Kimi” has to be capitalized because it’s the first word
in the name of a work. The word “ni” has to be lowercase because it’s a
particle.

That being said, I don’t actually know Japanese, and I used two
different online translators to generate the romaji. The only thing that
I tweaked was the capitalization. For all I know, the romanization could
be completely incorrect.

Later, I ended up doing the same thing for 浪漫飛行.

[1]: <https://japanese.stackexchange.com/a/18178/63845>
[2]: <https://www.loc.gov/catdir/cpso/romanization/japanese.pdf>