Mention that the build tool can take forever
If someone stumbles upon this repo and tries to build this game from
source, there’s a good chance that they’ll be confused by the build
tool. It can take a significant amount of time for the build tool to
produce any output at all. When I see that a command is running but
isn’t giving me any feedback, I get concerned that command isn’t
working. Hopefully, people will be less likely to get concerned now that
there’s a warning in the README.