Stop mypy pre-commit hook from failing
The mypy pre-commit hook was using a version of the reuse package that
was newer that the version of the reuse package in uv.lock. This was
causing the mypy pre-commit hook to fail. This change works around that
problem by locking the version of reuse that gets used by the mypy
pre-commit hook.

This change is a temporary workaround. Once uv.lock has been updated to
use a newer version of reuse, this commit should be reverted.