Warn users if they use a different version of Godot
The main motivation behind this change is to make it easier to create a
future commit. That future commit will change the revision of Nixpkgs
that this repo uses. Changing the revision of Nixpkgs that this repo
uses could potentially have the side effect of making this repo’s
devShell use the wrong version of Godot. This warning will help me avoid
that problem.