Make name of Python package more consistent
Before this change, the package sometimes called itself “build_tool”,
sometimes called itself “type-that-tune-build-tool” and sometimes called
itself “ttt-build-tool”. This change makes it so that it always calls
itself either “ttt_build_tool” or “ttt-build-tool”.

The main motivation behind this change is to make the Python package
easier to debug. I was working on a future commit, and I was really
struggling to figure out why my code wasn’t working. In the end, I
figured out that it wasn’t working because I thought that the package
was named “ttt-build-tool” but it was actually named
“type-that-tune-build-tool”. This change makes it harder to make a
mistake like that in the future.