Rename python_packages/ to python_distribution_packages/
I did this because the term “Python package” is ambiguous. It could
refer to a Python import package or a Python distribution package. This
repo uses the term “package” in other locations to refer to import
packages instead of distribution packages. The main motivation behind
this change is to (hopefully) make things a little less confusing.