Split the build tool into modules
The main motivation behind this change is to make it easier to create
future commits. Those future commits will allow the user to choose which
tasks they want to run when using the build tool and will add additional
tasks. It will be easier to create those two future commits if the code
is more organized.