Create prepare_godot_project task
The main motivation behind this change is to make it easier to create a
future commit. That future commit will add additional tasks. I want to
have a single task that runs all of the other tasks in the correct
order. The new prepare_godot_project task does exactly that.