Move tasks into their own subpackage
The main motivation behind this change is to make it easier to create a
future commit. That future commit will allow the user to choose what
task they want to run when running the build tool. Putting all of the
tasks in their own subpackage will allow me to search that subpackage
programmatically for the requested task.