Allow the user to choose a build task
The main motivation behind this change is to make it easier to create
future commits. Those future commits will add additional tasks to the
build tool. It will only make sense to run some of those tasks in
certain situations. By allowing the user to choose which build task they
want to run, the build tool will be able to only perform the tasks that
are applicable to whatever situation the user is currently in.