Make ttt-build-tool’s name consistent
Before this change, this repo would sometimes use different names when
referring to ttt-build-tool. This change makes it so that it’s always
called ttt-build-tool except in situations where hyphens in words are
not allowed. In those situations, this change makes it so that it’s
called ttt_build_tool.