Create ensure_editor_symlink task
The main motivation behind this change is to make it easier to create a
future commit. That future commit will add a task that will run the
Godot editor.

By default, the ensure_editor_symlink task builds a copy of the Godot
Engine editor from source. It does this in order lessen the number of
dependencies that the user needs in order to set up. The less
dependencies that a user needs to set up, the less effort it will take
for that user to build this project from source.