Add missing modules to ttt-build-tool distribution package
Before this change, there was a chance that installing the
ttt-build-tool distribution package would not give you all of the
modules that you need to run the ttt-build-tool. This change fixes that
issue.

The main motivation behind this change is to make it easier to create a
future commit. That future commit will add a Nix package for the
ttt-build-tool. The Nix package won’t work properly without this fix.