Create export_project task
In order to make it easier to use this new task on NixOS, a new
environment variable was added to the Nix devShell:
GODOT_EXPORT_TEMPLATES_PATH. Another variable, GODOT_EDITOR_PATH was
also added in order to match the GODOT_EXPORT_TEMPLATES_PATH variable.
The goal of these variables is to make it easier to export the game when
using the Nix devShell. You would use those two variables like this:

    uv run ttt-build-tool \
        --godot-editor-path "$GODOT_EDITOR_PATH" \
        --godot-export-templates-path "$GODOT_EXPORT_TEMPLATES_PATH" \
        --godot-export-preset linuxbsd_x86_64 \
        export_project

Without those two variables, you would have to put in more effort in
order to figure out the paths for the Godot editor and the export
templates.