Create Nix package for ttt-runtime-tool
A future commit is going to make the ttt-build-tool depend on the
ttt-runtime-tool. When I make that commit, I’m going to have to update
the ttt-build-tool Nix package so that it depends on a Nix package for
the ttt-runtime-tool. I’m creating the ttt-runtime-tool Nix package now
so that I don’t have to do it later.