Store settings in a build configuration file
Previous version of the ttt-build-tool made you specify all of the build
configuration using command-line options. That was kind of tedious. I
had considered telling users to just always create a batch file or a
shell script in order to make it less tedious, but that would require
platform-specific instructions. Build configuration files are platform
neutral.

I chose TOML for the build configuration file format because TOML is
easy to write. TOML can also be programmatically edited using toml-cli
[1]. It’s important to have a tool for programmatically generating TOML
because people creating packages for Type That Tune may need to change
build configurations conditionally. For example, the source file for a
particular package might need to change part of the build configuration
depending on which architecture the package is being built on.

I considered choosing JSON for the build configuration file format, but
I decided against it because JSON doesn’t support comments.

[1]: <https://github.com/gnprice/toml-cli>