Make ttt_build_tool respect Cache-Control headers
I don’t really know if this will have much of an affect on anything. I’m
just doing it because I think that it’s better to trust that the server
operator knows what they’re doing and that there’s a good reason why
they’re telling you to handle the cache in a certain way.