Create guide for updating pinned dependencies
Hopefully, this will help prevent me from forgetting to update things.
At the moment, the guide isn’t perfect (there is still some manual work
that has to be done that isn’t documented in the guide), but it’s better
than having nothing.

This is related to #18, but it doesn’t actually implement that feature
request.