Make build tool tell the user when tasks are run
Sometimes, the ttt-build-tool takes a very long time to complete. Before
this change, it would often take a very long time before the
ttt-build-tool printed anything.

This change makes it so that the ttt-build-tool prints things sooner and
more often. Hopefully, this change will help make it clear that the
ttt-build-tool is indeed making progress, even if it takes a really long
time to finish.

Closes #19.