Make Godot project check for the correct version
The Godot project will check to see what version of the Godot Engine you
used and emit a warning if the wrong version of Godot was used.
Unfortunately, a mistake was made, and the warning would be emitted even
if the user was using the correct version of Godot. This change fixes
that problem.

Fixes #37.