Create copies using copy.copy
When someone sees copy.copy(whatever), they know that it’s going to
create a copy of whatever. When someone sees dict(whatever), they may or
may not know that it’s going to crate a copy of whatever.

This change makes it so that we use copy.copy instead of dict when
creating copies. Hopefully, this will make the code easier to
understand.