Always use LFs, even on Windows
This repo’s EditorConfig file says to always use LFs for newlines, no
matter what platform the user is currently using. Before this change, if
you cloned this repo on Windows, then there was a chance that Git would
replace the LFs with CRLFs. This change makes sure that Git won’t try
and contradict the line ending type specified in this repo’s
EditorConfig.