Make sure the current working drive is changed
Before this change, users who were following the build instructions on
Windows and using the Windows Command Shell might have ended up
encountering a problem. Specifically, if the drive letter for the path
to their local clone of the type-that-tune repo is different than the
inital drive letter for a new Windows Command Shell session on their
system, then they would end up running uv in the wrong directory. This
is because (in the Windows Command Shell) the CD command doesn’t change
the current working drive letter by default. It only changes the current
working directory for a particular drive letter. For example, if the
user runs these commands…

	C:
	CD C:\Example directory 1
	CD D:\Example directory 2

…then they will still be in C:\Example directory 1. When they do switch
over to their D: drive, they’ll be in D:\Example directory 2, but they
still have to explictly switch to that drive afterwards.

This change ensures that users will not encounter that problem.