Make sure users can create symlinks
Before this change, if someone tried to follow the build instructions
on Windows, then there was a chance that they would encounter a problem.
Specifically, the ttt-build-tool will try to create a symlink and
there’s a chance that the user that’s running the ttt-build-tool doesn’t
have permission to create symlinks.

This change prevents that potential problem by having Windows users
enable Developer Mode. When Developer Mode is enabled, regular users are
allowed to create symlinks [1].

[1]: <https://neacsu.net/posts/win_symlinks/#3-enable-developer-mode>