Fix processor check on Windows
Before this change, if you tried to follow the README’s build
instructions on an x86_64 version of Windows, then you would get this
warning:

> UserWarning: The ttt-build-tool doesn’t know how to download a copy of
> the Godot Engine editor for your current platform.

That warning was being emitted because platform.machine() returns
"AMD64" on Windows. Before this change, I had assumed that it would
return "x86_64" on Windows because it returns "x86_64" on Linux.

This change fixes that warning.